feat: Add fetchRole function to user API
This commit adds a new function, fetchRole, to the user API in order to fetch the role of a user based on their UID. This function makes an HTTP GET request to the server and returns the role data from the response. If an error occurs during the request, null is returned.